Major changes to all files made.
Makefile moved to src/parser folder to overcome the dependencies problems. The idea
is to have a separate Makefile for every subfolder in src/ and to have a master
Makefile that will handle them all (TODO). Modified main.ml to be more readable until
we fix the Parser.mly which we modifie by adding actions-types and precedences to the
grammar. Modified Lexer.mli because had a syntax error and finally removed the token
type from Lexer.mll because we added those definitions in the Parser.mly file.